Several minor changes
- Moved ripple settings to radial grid instead of distribution functions
- Made it possible to give dB_B as an array of spatiotemporally uniform
  modes.
- Renamed 'IonInterpolator1D' to 'MultiInterpolator1D', which better
  reflects its potential usage.
