+++ src/MultiInterpolator1D.cpp
@@ -5,7 +5,7 @@
  * ion charge state density profiles in time.
  */
 
-#include "DREAM/IonInterpolator1D.hpp"
+#include "DREAM/MultiInterpolator1D.hpp"
 
 
 using namespace DREAM;
@@ -14,9 +14,9 @@ using namespace DREAM;
 /**
  * Constructor.
  */
-IonInterpolator1D::IonInterpolator1D(
+MultiInterpolator1D::MultiInterpolator1D(
     const len_t nZ0, const len_t nt, const len_t nr,
-    const real_t *t, const real_t *densities,
+    const real_t *t, const real_t *x,
     enum FVM::Interpolator1D::interp_method meth
 ) {
     this->nZ0 = nZ0;
@@ -24,12 +24,12 @@ IonInterpolator1D::IonInterpolator1D(
     this->nr  = nr;
     this->interps = new FVM::Interpolator1D*[nZ0];
 
-    this->t         = t;
-    this->densities = densities;
+    this->t = t;
+    this->x = x;
 
     for (len_t i = 0; i < nZ0; i++) {
         this->interps[i] = new FVM::Interpolator1D(
-            nt, nr, t, densities+(i*nt*nr),  meth
+            nt, nr, t, x+(i*nt*nr),  meth
         );
     }
 }
@@ -37,24 +37,24 @@ IonInterpolator1D::IonInterpolator1D(
 /**
  * Destrutor.
  */
-IonInterpolator1D::~IonInterpolator1D() {
+MultiInterpolator1D::~MultiInterpolator1D() {
     for (len_t i = 0; i < this->nZ0; i++)
         delete this->interps[i];
 
     delete [] this->interps;
 
     delete [] this->t;
-    delete [] this->densities;
+    delete [] this->x;
 }
 
 /**
- * Evaluate the radial density of the given ion
- * charge state, at the given time.
+ * Evaluate the radial density of the given parameter
+ * index, at the given time.
  *
- * iZ0:  Ion charge state index to evaluate radial density for.
- * t:    Time at which to evaluate the density profile.
+ * n: Index to evaluate radial density for.
+ * t: Time at which to evaluate the density profile.
  */
-const real_t *IonInterpolator1D::Eval(const len_t iZ0, const real_t t) {
-    return this->interps[iZ0]->Eval(t);
+const real_t *MultiInterpolator1D::Eval(const len_t n, const real_t t) {
+    return this->interps[n]->Eval(t);
 }
