+++ src/OtherQuantityHandler.cpp
@@ -279,16 +279,15 @@ void OtherQuantityHandler::DefineQuantities() {
     // Magnetic ripple resonant momentum
     if (tracked_terms->f_hot_ripple_Dxx != nullptr) {
         len_t nModes = tracked_terms->f_hot_ripple_Dxx->GetNumberOfModes();
-        DEF_FL("fluid/f_hot_ripple_m", "Magnetic ripple poloidal mode number for f_hot", qd->Store(this->tracked_terms->f_hot_ripple_Dxx->GetPoloidalModeNumbers()););
-        DEF_FL("fluid/f_hot_ripple_n", "Magnetic ripple toroidal mode number for f_hot", qd->Store(this->tracked_terms->f_hot_ripple_Dxx->GetToroidalModeNumbers()););
         DEF_FL_MUL("fluid/f_hot_ripple_pmn", nModes, "Magnetic ripple resonant momentum for f_hot [mc]", qd->Store(this->tracked_terms->f_hot_ripple_Dxx->GetResonantMomentum()[0]););
     }
     if (tracked_terms->f_re_ripple_Dxx != nullptr) {
         len_t nModes = tracked_terms->f_re_ripple_Dxx->GetNumberOfModes();
-        DEF_FL("fluid/f_re_ripple_m", "Magnetic ripple poloidal mode number for f_re", qd->Store(this->tracked_terms->f_re_ripple_Dxx->GetPoloidalModeNumbers()););
-        DEF_FL("fluid/f_re_ripple_n", "Magnetic ripple toroidal mode number for f_re", qd->Store(this->tracked_terms->f_re_ripple_Dxx->GetToroidalModeNumbers()););
         DEF_FL_MUL("fluid/f_re_ripple_pmn", nModes, "Magnetic ripple resonant momentum for f_re [mc]", qd->Store(this->tracked_terms->f_re_ripple_Dxx->GetResonantMomentum()[0]););
     }
+
+    DEF_FL("fluid/ripple_m", "Magnetic ripple poloidal mode number for f_hot", qd->Store(this->tracked_terms->f_hot_ripple_Dxx->GetPoloidalModeNumbers()););
+    DEF_FL("fluid/ripple_n", "Magnetic ripple toroidal mode number for f_hot", qd->Store(this->tracked_terms->f_hot_ripple_Dxx->GetToroidalModeNumbers()););
 
     // hottail/...
     DEF_HT_F1("hottail/nu_s_f1", "Slowing down frequency (on p1 flux grid) [s^-1]", qd->Store(nr_ht,   (n1_ht+1)*n2_ht, this->cqtyHottail->GetNuS()->GetValue_f1()););
@@ -340,8 +339,7 @@ void OtherQuantityHandler::DefineQuantities() {
     }
 
     this->groups["ripple"] = {
-        "fluid/f_hot_ripple_m", "fluid/f_hot_ripple_n", "fluid/f_hot_ripple_pmn",
-        "fluid/f_re_ripple_m", "fluid/f_re_ripple_n", "fluid/f_re_ripple_pmn"
+        "fluid/ripple_m", "fluid/ripple_n", "fluid/f_hot_ripple_pmn", "fluid/f_re_ripple_pmn"
     };
 
     this->groups["transport"] = {
